third_party/blink/web_tests/platform/mac-mac10.14/external/wpt/html/browsers/windows/clear-window-name.https-expected.txt

This is a testharness.js-based test.
PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
FAIL Window.name is reset at the first cross-origin navigation with noopener assert_equals: expected "" but got "226f125c-5ecd-4b28-b1fd-34f4c31d59ff"
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
FAIL Window.name is reset at the first first cross-origin navigation with window.opener set to null assert_equals: expected "" but got "b56be3ce-f279-4df4-aae9-c6d1d1807c15"
PASS Window.name is set by the window
FAIL Window.name is reset at the first cross-origin navigation assert_equals: expected "" but got "d865e2b1-1e5c-46d9-b64c-413f9f967973"
Harness: the test ran to completion.
